+++ docs/releasenotes.rst
@@ -17,6 +17,7 @@ Commits by Gabriel Beckers.
 - improved audio codec handling when extracting audio (best muxer chosen automatically)
 - refactored ffmpeg module: a.o. better error handling
 - implemented a VideoFile class
+- improved stream awareness (files can have multiple video and audio streams)
 
 
 Version 0.5